feat: Disable any logging from go-client's klog
This PR disables warnings from go-client. Unfortunately this is not
sufficient to also supress the warning about deprecated auth plugins,
therefore the need for klog output code too.

But in general direct logging from 3rd-party libs is undesirable, and
this configures K8S's go-client klog to discard any logs.

I've added some tests to cover some unexpected 3rd-party output, but as
this depends on the environment and kubeconfig config, incl. providers,
and is noteasily reproducible, this should be covered by integration
test, as the current test do not fully cover this. Opened #406.

fixes #402